% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_helpers.R
\name{complex_heatmap}
\alias{complex_heatmap}
\title{Heatmap with p-value significance level labels.}
\usage{
complex_heatmap(
  zscores_df,
  pvalues_df,
  row_title = "",
  column_title = "",
  legend_title = "",
  fig_width = 10,
  fig_height = NULL
)
}
\arguments{
\item{zscores_df}{`data.frame`. Data containing fold changes or z-scores.}

\item{pvalues_df}{`data.frame`. Data containing p-values or q-values.}

\item{row_title}{`string`. The row title.}

\item{column_title}{`string`. The column title.}

\item{legend_title}{`string`. The key title.}

\item{fig_width}{`numeric`. The figure width (inches).}

\item{fig_height}{`numeric`. The figure height (inches). Default resizes
dynamically according to number of rows.}
}
\value{
`Large Heatmap`. See \code{\link{ComplexHeatmap::Heatmap}}
}
\description{
Heatmap with p-value significance level labels.
}
