% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_helpers.R
\name{fix_combiPeptData}
\alias{fix_combiPeptData}
\title{Fix incorrect gene names in Pescal Excel file}
\usage{
fix_combiPeptData(
  df.combi = df.combi,
  organism = organism_dbs,
  pescal.xlsx = pescal.output.file,
  fixed.xlsx = df.combi.fixed.xlsx,
  save.xlsx = TRUE
)
}
\arguments{
\item{df.combi}{\code{\link{data.frame}}. combiPeptData.}

\item{organism}{String. One of 'human', 'mouse', 'rat', or 'pig'. Selects
database of Uniprot names.}

\item{pescal.xlsx}{String. The path ("directory/name") of the Pescal Excel
output.}

\item{fixed.xlsx}{String. Name for the fixed output Excel file.}
}
\value{
\code{\link{data.frame}}. The corrected combiPeptData. Additionally,
saves a copy of the Pescal Excel file with the corrected combiPeptData sheet
(with the name of `fixed.xlsx`).
}
\description{
Fixes incorrect data of pre-processed mass spectrometry Excel file:
combiPeptData columns 25 and 29.
}
