% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_helpers.R
\name{installer}
\alias{installer}
\title{Install multiple missing packages and load.}
\usage{
installer(
  pkg_cran = NULL,
  pkg_git = NULL,
  pkg_biocmanager = NULL,
  update_all = FALSE,
  contriburl = "https://cloud.r-project.org/"
)
}
\arguments{
\item{pkg_cran}{A character vector of CRAN package names.}

\item{pkg_git}{A character vector of GitHub repositories, each in the
format: `"Username/repository"`.}

\item{pkg_biocmanager}{A character vector of Bioconductor package names.}

\item{update_all}{A logical. Default `FALSE` only installs missing packages.
If `TRUE`, the current versions of all packages will be installed.}

\item{contriburl}{A character string of the repository mirror URL. See
\code{\link{install.packages}}.}
}
\value{
No return value. Installs missing/all packages and loads all arguments.
}
\description{
Given vectors of CRAN packages, GitHub repositories, and Bioconductor
  packages, checks which are not installed and installs only the missing
  packages/repositories. If `update_all` is `TRUE`, then all packages are
  installed instead, even packages already installed. The function then
  loads all supplied packages.
}
\seealso{
\itemize{
  \code{\link{install.packages}}
  \code{\link{installed.packages}}
  \code{\link{devtools}}
  \code{\link{devtools::install_github}}
  \code{\link{BiocManager}}
  \code{\link{BiocManager::install}}
}
}
