% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_helpers.R
\name{mergeDTs}
\alias{mergeDTs}
\title{Merges multiple \code{\link{data.table}}s.}
\usage{
mergeDTs(dt_list, by = NULL, all = TRUE, sort = FALSE)
}
\arguments{
\item{dt_list}{\code{\link{list}} of \code{\link{data.table}}s.}

\item{by}{String/vector of strings. The column(s) to merge on.}

\item{all}{Logical. Default `TRUE` is an outer join; `FALSE` is an inner join.}

\item{sort}{Logical. To sort on the `by` columns. Default `FALSE`.}
}
\value{
A \code{\link{data.table}}. The result of the merge.
}
\description{
Merges multiple \code{\link{data.table}}s.
}
\examples{
DT <- data.table(x=rep(c("b","a","c"),each=3), y=c(1,3,6), v=1:9)
X <- data.table(x=c("c","b"), v=8:7, foo=c(4,2))

multi_DT <- mergeDTs(
  list(DT, X),
  by = "x",
  all = TRUE  # Outer join
)
}
