% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_helpers.R
\name{myHeatMap_With_PValues}
\alias{myHeatMap_With_PValues}
\title{Heatmap with p-value significance level labels.}
\usage{
myHeatMap_With_PValues(df.zcr, df.pval, mytitle = "", key.title = "log2Fold")
}
\arguments{
\item{df.zcr}{A `data.frame` with fold changes or z-scores}

\item{df.pval}{A `data.frame` with p-values or q-values}

\item{mytitle}{A `string` for the heatmap title.}

\item{key.title}{A `string` for the key title, defaults to "log2Fold".}
}
\value{
A \code{\link{gplots::heatmap.2}} heatmap.
}
\description{
Heatmap with p-value significance level labels.
}
