% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_helpers.R
\name{remove_outlier_samples_from_dataset_edit}
\alias{remove_outlier_samples_from_dataset_edit}
\title{Remove outliers from a dataframe.}
\usage{
remove_outlier_samples_from_dataset_edit(df, plot = TRUE)
}
\arguments{
\item{df}{A `data.frame`.}

\item{plot}{Boolean. If TRUE (default) then the boxplot is produced. If
FALSE then no plot is produced.}
}
\value{
Returns the given dataset with columns identified as outliers
  removed. If no outliers are present in the dataset, then the original
  dataset is returned.
}
\description{
Takes a dataframe in wide format, calculates medians for each numeric
  column, and uses `boxplot()$out` values to identify outliers for removal.
}
\seealso{
\code{\link{boxplot}}
}
