% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_helpers.R
\name{umap_and_plot}
\alias{umap_and_plot}
\title{Uniform Manifold Approximation and Projection (UMAP)}
\usage{
umap_and_plot(
  input,
  label = NULL,
  size = 3,
  max.overlaps = getOption("ggrepel.max.overlaps", default = 10),
  title = "UMAP",
  guide = FALSE
)
}
\arguments{
\item{input}{\code{\link{data.frame}} or \code{\link{matrix}}-like object.
Contains numeric data.}

\item{label}{`'numeric'`, `'character'`, or `NULL`. Default `NULL` uses
\code{\link{rownames}}, else labels are taken from the specified column number
(`'numeric'`) or name (`'character'`) and separated from the numeric data.}

\item{size}{`'numeric'`. Label text size, default = `3`.}

\item{max.overlaps}{`numeric`. \code{\link{ggrepel::geom_label_repel}} `max.overlaps`
parameter to exclude text labels that overlap too many items; default = `10`.}

\item{title}{`'character'`. The title of the plot.}

\item{guide}{`bool`. Plots the legend if `TRUE`, default `FALSE` does not.}
}
\value{
\code{\link{list}}. A list of two elements:
\enumerate{
  \item \code{\link{umap::umap}} object. Containing at least a component with an
  embedding and a component with configuration settings.
  \item \code{\link{ggplot2::ggplot}} object. Plot of the UMAP result.
}
}
\description{
Performs Uniform Manifold Approximation and Projection (UMAP) dimensionality
reduction on a matrix-like object using the \code{\link{umap::umap}} package.
}
\section{Notes}{

The UMAP algorithm is stochastic, so repeated executions will yield different results.
}

\section{Warnings}{

If output:
\preformatted{
`Warning: ggrepel: _ unlabelled data points (too many overlaps). Consider increasing max.overlaps
}
Resolve by any of the following:
\enumerate{
  \item Increase plot viewer size and/or increase chunk options `fig.width` and `fig.height`.
  \item Decrease argument of `size`.
  \item Increase argument of `max.overlaps`.
}
}

\seealso{
\itemize{
  \item \code{\link{umap::umap}}
}
}
