% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_helpers.R
\name{update_protools}
\alias{update_protools}
\title{Update to the latest release of the `protools2` package.}
\usage{
update_protools(
  url = "https://api.github.com/repos/iibadshah/protools2/releases/latest"
)
}
\arguments{
\item{url}{String. URL pointing to GitHub release.}
}
\value{
None.
}
\description{
Update to the latest release of the `protools2` package.
}
\examples{
# As it installs the `devtools2` package, it must not be loaded. Therefore,
# call like so:
devtools2::update_protools()
}
